git-svn-id: https://svn.apache.org/repos/asf/sis/trunk@1554593 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-referencing/src/main/java/org/apache/sis/io/wkt/WKTFormat.java
@@ -27,5 +27,5 @@ public class WKTFormat {
     /**
      * The default indentation value.
      */
-    static final int DEFAULT_INDENTATION = 2;
+    static final byte DEFAULT_INDENTATION = 2;
 }
